move track panel bar icons to sidebar folder
