+++ src/ensembl/src/content/app/browser/track-panel/track-panel-bar/trackPanelBarConfig.ts
@@ -1,14 +1,9 @@
-import { ReactComponent as bookmarkIcon } from 'static/img/track-panel/bookmark.svg';
-
-import { ReactComponent as downloadIcon } from 'static/img/track-panel/download.svg';
-
-import { ReactComponent as searchIcon } from 'static/img/track-panel/search.svg';
-
-import { ReactComponent as personalDataIcon } from 'static/img/track-panel/own-data.svg';
-
-import { ReactComponent as shareIcon } from 'static/img/track-panel/share.svg';
-
-import { ReactComponent as tracksManagerIcon } from 'static/img/track-panel/tracks-manager.svg';
+import { ReactComponent as bookmarkIcon } from 'static/img/sidebar/bookmark.svg';
+import { ReactComponent as downloadIcon } from 'static/img/sidebar/download.svg';
+import { ReactComponent as searchIcon } from 'static/img/sidebar/search.svg';
+import { ReactComponent as personalDataIcon } from 'static/img/sidebar/own-data.svg';
+import { ReactComponent as shareIcon } from 'static/img/sidebar/share.svg';
+import { ReactComponent as tracksManagerIcon } from 'static/img/sidebar/tracks-manager.svg';
 
 export type TrackPanelBarItem = {
   description: string;
